Revert "Revert "Revert "Revert "Remove attempt at citations.""""
We cannot control where section header is placed. Pandoc adds it inside post's
`$body$`, so there is no way to control this via templates. However, there is
also no way to control bibliography layout from within the CSL file. the only
workaround, of adding an empty `References` section at the end of the article
is bad because it requires manual work and also needs HTML escape to insert
proper heading (which means that when template change it will break silently).

Pandoc only supports XML CSL, so we cannot reuse the LaTeX ones anyway.

This reverts commit 3634ee8d4fefbe019b87b6468705bc1334ccf929.