Try a new version of `shell.nix` to update GHC.
This version attempts to add `diagrams` (and `palette` by the time this got
checked in) to the "system GHC" that is available to the Nix shell. The GHC
controlled by Cabal has no issue, but `diagrams-builder` uses the system GHC so
it needs to be able to find the needed packages.

Signed-off-by: Mihai Maruseac <mihai.maruseac@gmail.com>